requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cheroot==8.5.1
CherryPy==18.6.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
elasticsearch6==6.8.1
Flask==1.1.2
hyperlink==20.0.1
idna==2.10
importlib-resources==4.1.1
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.1.0
jaraco.functools==3.0.1
jaraco.text==3.4.0
Jinja2==2.11.2
jmespath==0.10.0
kafka-python==2.0.2
lxml==4.6.2
MarkupSafe==1.1.1
more-itertools==8.6.0
parsel==1.6.0
Paste==3.5.0
portend==2.7.0
Protego==0.1.16
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyspark==3.0.1
pytz==2020.5
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
tempora==4.0.1
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
Werkzeug==1.0.1
zc.lockfile==2.0
zipp==3.4.0
zope.interface==5.2.0
